k1c: Add support for k1c architecture
Summary:
Implement strace support for k1c. Thanks to REGSET, there is only a few
things to implement to have it working. The current implementation is
simply an application of the ABI since syscalls are seen as standard
procedure calls on k1c.

Test Plan:
Ran strace in kernel (see D1138) and checked output is correct to
expected strace values.

Reviewers: O51 Linux Coolidge!

Differential Revision: https://phab.kalray.eu/D1144