k1c: remove r6 from clobber registers list
Summary:
GCC complains that r6 is also in the clobber list while being in the
input list. To be consistent with other register r0 -> r7, set it as
input/ouput using +r to say it will be cloberred too.

Test Plan: Build tests and verify that syscall are correctly executed.

Differential Revision: https://phab.kalray.eu/D3088